+++ demo/AEPCommerceDemoApp/AEPCommerceDemoApp/Product.swift
@@ -15,7 +15,7 @@ import Foundation
 struct Product: Equatable {
     var productData : ProductData
     var quantity : Int
-    var subtotal: Float { return productData.price * Float(quantity) }
+    var subtotal: Double { return productData.price * Double(quantity) }
 }
 
 extension Product {
